Refactor tests and remove stale tests / warnings
Certain tests (related to callbacks on membership that we moved
to the controller, for example) were marked pending and needed to be
simply removed.  Also a GroupStat test was skipped and needed
refactoring anyway to use stubs instead of a convoluted timecop
solution.

Also skylight was causing warnings so I removed it. We don't
use it anyway.